Atualizações dos parametors de produtos e acessos + atualização dos caminhos na main

+++ index/index.h
@@ -2,13 +2,13 @@
 #define INDEX_H
 
 typedef struct {
-    int key;         // Campo chave (ID do produto ou ID da sessão)
-    long offset;     // Deslocamento no arquivo de dados
+    unsigned long long key;      // Campo chave (ID do produto)
+    long offset;                 // Deslocamento no arquivo de dados
 } IndexEntry;
 
 void create_index(const char *data_filename, const char *index_filename);
-IndexEntry* binary_search_index(const char *index_filename, int target_key);
-void search_with_index(const char *data_filename, const char *index_filename, int target_key);
-void remove_product(const char *data_filename, const char *index_filename, int target_id);
+IndexEntry* binary_search_index(const char *index_filename, unsigned long long target_key);
+void search_with_index(const char *data_filename, const char *index_filename, unsigned long long target_key);
+void remove_product(const char *data_filename, const char *index_filename, unsigned long long target_id);
 
 #endif
